requirements.txt
# FastAPI & ASync Server
fastapi==0.110.0
uvicorn[standard]==0.27.1

# AI Orchestration
langchain==0.1.13
langgraph==0.0.31
langchain-openai==0.1.1
langchain-core==0.1.33

# Database & Vectors
SQLAlchemy[asyncio]==2.0.29
sqlmodel==0.0.16
pgvector==0.2.5
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Document & RAG Utilities
pypdf==4.1.0
tiktoken==0.6.0
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# Optional Tooling/Tests
pytest==8.1.1
